Finalizando o edit e delete da person

+++ src/app/modules/person/person-list/person-list.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { ResponsePageable } from 'src/app/components/shared/model/reponsePageable.model';
+import Swal from 'sweetalert2';
 import { Person } from '../interface/person.interface';
 import { PersonService } from '../service/person.service';
 
@@ -12,8 +13,9 @@ import { PersonService } from '../service/person.service';
 export class PersonListComponent implements OnInit {
   rule!: String;
   listPerson: any
-  list: any
+  list!: ResponsePageable
   ruleTitle: any
+  totalPages!: any
 
   constructor(
     private route: ActivatedRoute,
@@ -37,13 +39,33 @@ export class PersonListComponent implements OnInit {
   }
 
   async getListPerson(page: number) {
-    this.list = await this.personService.getPersonListByRule(this.rule, 0).toPromise().then(response => response)
+    this.list = await this.personService.getPersonListByRule(this.rule, page).toPromise().then(response => response)
     this.listPerson = this.list.content
+    this.totalPages = Array(this.list.totalPages).map((x,i)=>i);
   }
 
   deletePersonById(id: any) {
-    this.personService.deleteById(id)
-    location.reload()
+    Swal.fire({
+      title: 'Você tem certeza?',
+      text: "Você não será capaz de reverter isso!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Sim, excluir',
+      cancelButtonText: 'Cancelar',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.personService.deleteById(id)
+        Swal.fire(
+          'Excluído!',
+          'Cadastro deletado com sucesso.',
+          'success'
+        )
+        setTimeout(() => {
+          location.reload()
+        }, 500)
+      }
+    })
   }
-
 }
